[PR Merge] Core Integration
[Units] hw/units contains the rtl files for custom ips. The directory name must match the ip name in hw/xilinx/ips/common.
Specifically, it must start with custom_; Use the custom_top_wrapper.sv file in the hw/units/template directory
and fill the fetch_sources.sh script to retrieve the rtl sources for the custom IP.
[Flow] hw/xilinx/Makefile: Duplicated variables for XILINXIP_LIST and derived for Custom IPs (CUSTOM*_IP_LIST).
[Flow] hw/xilinx/Makefile: Changed the IP_LIST variable to accomodate the new project structure path (see below)
[Flow] hw/xilinx/Makefile: Changed XILINX_IP_NAMES to IP_NAMES
[Flow] Integration for the custom IP flow; the package_ip.tcl can be used to package new custom IPs
[UninaSoC] Integration of the new RVM socket module. It exposes two master AXI interfaces for instruction and data.
It can support every processor compatible with either the MEM or the AXI interface. Currently, the repo already has
a fetch_sources.sh scripts for cv32e40p and picorv32.
[Documentation] Adding a documentation (README.md) for the custom IP creation process.