[verify/plic_and_interrupts]  Merging branch after verification and refactoring (#42)
PLIC and Peripherals integration
- [hw/units] Includ Custom AXI PLIC peripheral
- [hw/xilinx/ips] Add AXI Timers (TIM0, TIM1), GPIO_in (embedded only)
- [hw/xilinx] Add PLIC peripheral mapped to the main bus, TIM0, TIM1 and
GPIO mapped to periphral bus
- [hw/xilinx] Peripheral bus is now generating interrupts
- [sw/SoC] Add PLIC software example using tnterrupts from GPIO_in and
TIM0
- [hw/xilinx] PLIC forwards a single interrupt line, which is
concatenated in a 32-bit word to the socket
- [cores] CV32E40P is supported
- [cores] PICORV32 does not support standard RISC-V interrupts.
- [cores] Microblaze does not support standard RISC-V interrupts.
Apparentely, the microblaze-V does not correctly handle interrupts
(probably a bug of the amd IP from Vivado 2024.2). It correctly stacks
the MPP bit, resets the mstatus.MIE and sets the bit in MIP, but does
not jump to the vector table.

---------

Co-authored-by: Valedd99 <didomenico.valerio@virgilio.it>