[feature/config/cdc] Configuration-based CDC (#91)
* [config] update the config flow to include clock domains (MAIN_CLOCK_DOMAIN and RANGE_CLOCK_DOMAINS)
* [config/doc] Update doc
* [cdc] add <SLAVE_NAME>_HAS_CLOCK_DOMAIN as Verilog defines 
* [cdc] add conditional CDC bridge for PBUS
* [cdc] add reset synchronizers in sys_master 
* [embedded][hw/xilinx/ips] Propagate PBUS frequency as config parameter to AXI-lite UART synthesis

NOTE: In the hpc profile, the configuration with the main_clock_domain at 100MHz and the pbus_clock_domain at 250MHz synthetizes with "Vivado Default Synthesis" and "Vivado Default Implementation". The configuration with the main_clock_domain at 100MHz and the pbus_clock_domain at 50MHz does not meet the timing, you should add retiming and use "explore" and "aggressive explore" strategies in place and route.

---------

Co-authored-by: Vincenzo Maisto <vincenzo.maisto2@unina.it>